requirements .txt
streamlit
opencv-python
pandas
Pillow
scikit-learn
facenet-pytorch
torch
torchvision